feat: split api group documentation for DPS module
