+++ data-protocols/data-plane-signaling/data-plane-signaling-core/src/main/java/org/eclipse/edc/signaling/port/api/management/DataPlaneRegistrationApiV5.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2025 Think-it GmbH
+ *  Copyright (c) 2026 Think-it GmbH
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.signaling.port.api;
+package org.eclipse.edc.signaling.port.api.management.v5;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.Operation;
@@ -24,15 +24,16 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 import org.eclipse.edc.api.model.ApiCoreSchema;
 import org.eclipse.edc.signaling.domain.DataPlaneRegistrationMessage;
 
 import static jakarta.ws.rs.HttpMethod.DELETE;
 import static jakarta.ws.rs.HttpMethod.PUT;
 
-@OpenAPIDefinition(info = @Info(version = "v4"))
-@Tag(name = "Dataplane Signaling Registration v4")
-public interface DataPlaneRegistrationApiV4 {
+@OpenAPIDefinition(info = @Info(version = "v5"))
+@Tag(name = "Dataplane Signaling Registration v5beta")
+public interface DataPlaneRegistrationApiV5 {
 
     @Operation(
             method = PUT,
@@ -46,7 +47,7 @@ public interface DataPlaneRegistrationApiV4 {
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiCoreSchema.ApiErrorDetailSchema.class))))
             }
     )
-    Response register(DataPlaneRegistrationMessage registration);
+    Response registerV5(String participantContextId, DataPlaneRegistrationMessage registration, SecurityContext securityContext);
 
 
     @Operation(
@@ -58,6 +59,6 @@ public interface DataPlaneRegistrationApiV4 {
                             content = @Content(array = @ArraySchema(schema = @Schema(implementation = ApiCoreSchema.ApiErrorDetailSchema.class))))
             }
     )
-    Response delete(String dataplaneId);
+    Response deleteV5(String participantContextId, String dataplaneId, SecurityContext securityContext);
 
 }
