+++ data-protocols/data-plane-signaling/data-plane-signaling-core/src/main/resources/signaling-api-version.json
@@ -2,6 +2,6 @@
   {
     "version": "1.0.0-alpha",
     "urlPath": "/v1",
-    "lastUpdated": "2026-04-01T14:00:01Z"
+    "lastUpdated": "2026-04-17T14:00:01Z"
   }
 ]
