--- data-protocols/data-plane-signaling/src/main/resources/signaling-api-version.json
@@ -1,7 +1,0 @@
-[
-  {
-    "version": "1.0.0-alpha",
-    "urlPath": "/v1",
-    "lastUpdated": "2026-04-16T15:00:01Z"
-  }
-]
