+++ data-protocols/dsp/dsp-2025/dsp-catalog-2025/dsp-catalog-http-api-2025/build.gradle.kts
@@ -43,6 +43,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api")
+        apiGroup("dsp-api")
     }
 }
