+++ data-protocols/dsp/dsp-2025/dsp-negotiation-2025/dsp-negotiation-http-api-2025/build.gradle.kts
@@ -42,6 +42,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api")
+        apiGroup("dsp-api")
     }
 }
