+++ data-protocols/dsp/dsp-version/dsp-version-http-api/build.gradle.kts
@@ -34,6 +34,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api")
+        apiGroup("dsp-api")
     }
 }
