+++ data-protocols/dsp/dsp-virtual/dsp-2025-virtual/dsp-negotiation-http-api-2025-virtual/build.gradle.kts
@@ -41,6 +41,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api-virtual")
+        apiGroup("dsp-api-virtual")
     }
 }
