+++ data-protocols/dsp/dsp-virtual/dsp-2025-virtual/dsp-transfer-process-http-api-2025-virtual/build.gradle.kts
@@ -42,6 +42,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api-virtual")
+        apiGroup("dsp-api-virtual")
     }
 }
