+++ data-protocols/dsp/dsp-virtual/dsp-metadata-http-api-virtual/build.gradle.kts
@@ -35,6 +35,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("dsp-api-virtual")
+        apiGroup("dsp-api-virtual")
     }
 }
