+++ extensions/common/api/api-observability/build.gradle.kts
@@ -30,7 +30,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("observability-api")
+        apiGroup("observability-api")
     }
 }
