+++ extensions/common/api/management-api-configuration/src/main/resources/management-api-version.json
@@ -8,7 +8,7 @@
   {
     "version": "4.0.0",
     "urlPath": "/v4",
-    "lastUpdated": "2026-04-09T09:43:01Z",
+    "lastUpdated": "2026-04-17T09:43:01Z",
     "maturity": "stable"
   },
   {
