+++ extensions/control-plane/api/control-plane-api/build.gradle.kts
@@ -38,7 +38,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("control-api")
+        apiGroup("control-api")
     }
 }
