+++ extensions/control-plane/api/management-api-v5/catalog-api-v5/build.gradle.kts
@@ -42,7 +42,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
