+++ extensions/control-plane/api/management-api-v5/cel-api-v5/build.gradle.kts
@@ -38,7 +38,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
