+++ extensions/control-plane/api/management-api-v5/contract-agreement-api-v5/build.gradle.kts
@@ -37,7 +37,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
