+++ extensions/control-plane/api/management-api-v5/contract-negotiation-api-v5/build.gradle.kts
@@ -35,7 +35,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
