+++ extensions/control-plane/api/management-api-v5/participant-context-config-api-v5/build.gradle.kts
@@ -39,7 +39,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
