+++ extensions/control-plane/api/management-api-v5/transfer-process-api-v5/build.gradle.kts
@@ -39,6 +39,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
