+++ extensions/control-plane/api/management-api/build.gradle.kts
@@ -28,5 +28,3 @@ dependencies {
     api(project(":extensions:control-plane:api:management-api:policy-definition-api"))
     api(project(":extensions:control-plane:api:management-api:transfer-process-api"))
 }
-
-
