+++ extensions/control-plane/api/management-api/catalog-api/build.gradle.kts
@@ -41,7 +41,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
