+++ extensions/control-plane/api/management-api/contract-definition-api/build.gradle.kts
@@ -42,7 +42,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
