+++ extensions/control-plane/api/management-api/edr-cache-api/build.gradle.kts
@@ -42,7 +42,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
