+++ extensions/control-plane/api/management-api/policy-definition-api/build.gradle.kts
@@ -50,7 +50,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
