+++ extensions/control-plane/api/management-api/transfer-process-api/build.gradle.kts
@@ -43,7 +43,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
