+++ extensions/data-plane-selector/data-plane-selector-api/build.gradle.kts
@@ -40,7 +40,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
