+++ extensions/data-plane-selector/data-plane-selector-control-api/build.gradle.kts
@@ -37,7 +37,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("control-api")
+        apiGroup("control-api")
     }
 }
