+++ extensions/data-plane/data-plane-provision-http/build.gradle.kts
@@ -36,7 +36,7 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
