+++ extensions/data-plane/data-plane-signaling/data-plane-signaling-api/build.gradle.kts
@@ -34,7 +34,7 @@ dependencies {
 }
 edcBuild {
     swagger {
-        apiGroup.set("control-api")
+        apiGroup("control-api")
     }
 }
