+++ extensions/data-plane/data-plane-signaling/data-plane-signaling-transform/build.gradle.kts
@@ -29,10 +29,3 @@ dependencies {
     testImplementation(project(":core:common:junit"))
     testImplementation(project(":core:common:lib:json-ld-lib"))
 }
-edcBuild {
-    swagger {
-        apiGroup.set("control-api")
-    }
-}
-
-
