+++ extensions/federated-catalog/api/federated-catalog-api/build.gradle.kts
@@ -50,6 +50,6 @@ dependencies {
 
 edcBuild {
     swagger {
-        apiGroup.set("management-api")
+        apiGroup("management-api")
     }
 }
