+++ gradle/libs.versions.toml
@@ -132,6 +132,6 @@ dsp-core = { module = "org.eclipse.dataspacetck.dsp:core", version.ref = "dcp-tc
 jackson = ["jackson-annotations", "jackson-databind"]
 
 [plugins]
-edc-build = { id = "org.eclipse.edc.edc-build", version = "1.4.0" }
+edc-build = { id = "org.eclipse.edc.edc-build", version = "1.5.1" }
 shadow = { id = "com.gradleup.shadow", version = "9.3.1" }
 swagger = { id = "io.swagger.core.v3.swagger-gradle-plugin", version.ref = "swagger" }
