+++ resources/openapi/signaling-api.version
@@ -1,1 +1,1 @@
-data-protocols/data-plane-signaling/src/main/resources/signaling-api-version.json
+data-protocols/data-plane-signaling/data-plane-signaling-core/src/main/resources/signaling-api-version.json
